Testing: update logic and improve modularity
