+++ tests/test_importer.py
@@ -5,167 +5,46 @@ Run with:
     pytest -v tests/test_importer.py
 """
 
-import zipfile
 import pytest
 from pathlib import Path
-from datetime import timezone  # kept for potential future use
 
-from opensak.db.database import init_db, get_session
+from opensak.db.database import get_session
 from opensak.db.models import Cache
 from opensak.importer import import_gpx, import_zip
 
-# ── Synthetic GPX matching real Groundspeak/PQ format ────────────────────────
+from tests.data import (
+    SAMPLE_GPX, SAMPLE_WPTS_GPX, EMPTY_GPX,
+    make_variant_gpx, make_gpx_with_inline_wpt, write_gpx, make_zip,
+)
 
-SAMPLE_GPX = """\
-<?xml version="1.0" encoding="utf-8"?>
-<gpx xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-     version="1.0"
-     creator="Groundspeak Pocket Query"
-     xmlns="http://www.topografix.com/GPX/1/0">
-  <n>Test PQ</n>
-  <time>2026-03-10T07:22:19Z</time>
-  <wpt lat="55.6761" lon="12.5683">
-    <time>2024-06-01T00:00:00</time>
-    <n>GC12345</n>
-    <desc>Test Traditional by TestOwner, Traditional Cache (2.0/3.0)</desc>
-    <url>https://coord.info/GC12345</url>
-    <urlname>Test Traditional</urlname>
-    <sym>Geocache</sym>
-    <type>Geocache|Traditional Cache</type>
-    <groundspeak:cache id="9999" archived="False" available="True"
-        xmlns:groundspeak="http://www.groundspeak.com/cache/1/0/1">
-      <groundspeak:name>Test Traditional</groundspeak:name>
-      <groundspeak:placed_by>TestOwner</groundspeak:placed_by>
-      <groundspeak:owner id="12345">TestOwner</groundspeak:owner>
-      <groundspeak:type>Traditional Cache</groundspeak:type>
-      <groundspeak:container>Small</groundspeak:container>
-      <groundspeak:attributes>
-        <groundspeak:attribute id="6" inc="1">Recommended for kids</groundspeak:attribute>
-        <groundspeak:attribute id="24" inc="0">Wheelchair accessible</groundspeak:attribute>
-      </groundspeak:attributes>
-      <groundspeak:difficulty>2.0</groundspeak:difficulty>
-      <groundspeak:terrain>3.0</groundspeak:terrain>
-      <groundspeak:country>Denmark</groundspeak:country>
-      <groundspeak:state>Zealand</groundspeak:state>
-      <groundspeak:short_description html="False">A short description.</groundspeak:short_description>
-      <groundspeak:long_description html="True">&lt;p&gt;A longer description.&lt;/p&gt;</groundspeak:long_description>
-      <groundspeak:encoded_hints>Under a rock.</groundspeak:encoded_hints>
-      <groundspeak:logs>
-        <groundspeak:log id="111">
-          <groundspeak:date>2026-01-15T10:30:00Z</groundspeak:date>
-          <groundspeak:type>Found it</groundspeak:type>
-          <groundspeak:finder id="999">Tester</groundspeak:finder>
-          <groundspeak:text encoded="False">TFTC! Great hide.</groundspeak:text>
-        </groundspeak:log>
-        <groundspeak:log id="112">
-          <groundspeak:date>2025-12-01T09:00:00Z</groundspeak:date>
-          <groundspeak:type>Didn't find it</groundspeak:type>
-          <groundspeak:finder id="888">Searcher</groundspeak:finder>
-          <groundspeak:text encoded="False">Could not find it.</groundspeak:text>
-        </groundspeak:log>
-      </groundspeak:logs>
-    </groundspeak:cache>
-  </wpt>
-  <wpt lat="56.1234" lon="10.5678">
-    <time>2023-03-15T00:00:00</time>
-    <n>GC99999</n>
-    <desc>Mystery Cache by Puzzler, Unknown Cache (4.0/2.5)</desc>
-    <url>https://coord.info/GC99999</url>
-    <urlname>Mystery Cache</urlname>
-    <sym>Geocache</sym>
-    <type>Geocache|Unknown Cache</type>
-    <groundspeak:cache id="8888" archived="False" available="True"
-        xmlns:groundspeak="http://www.groundspeak.com/cache/1/0/1">
-      <groundspeak:name>Mystery Cache</groundspeak:name>
-      <groundspeak:placed_by>Puzzler</groundspeak:placed_by>
-      <groundspeak:owner id="54321">Puzzler</groundspeak:owner>
-      <groundspeak:type>Unknown Cache</groundspeak:type>
-      <groundspeak:container>Micro</groundspeak:container>
-      <groundspeak:attributes/>
-      <groundspeak:difficulty>4.0</groundspeak:difficulty>
-      <groundspeak:terrain>2.5</groundspeak:terrain>
-      <groundspeak:country>Denmark</groundspeak:country>
-      <groundspeak:state>Region Midtjylland</groundspeak:state>
-      <groundspeak:short_description html="False"></groundspeak:short_description>
-      <groundspeak:long_description html="False">Solve the puzzle first.</groundspeak:long_description>
-      <groundspeak:encoded_hints>Check the tree.</groundspeak:encoded_hints>
-      <groundspeak:logs/>
-    </groundspeak:cache>
-  </wpt>
-</gpx>
-"""
-
-SAMPLE_WPTS_GPX = """\
-<?xml version="1.0" encoding="utf-8"?>
-<gpx xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-     version="1.0"
-     creator="Groundspeak, Inc."
-     xmlns="http://www.topografix.com/GPX/1/0">
-  <n>Waypoints for Cache Listings</n>
-  <wpt lat="55.6762" lon="12.5680">
-    <n>PK2345</n>
-    <desc>Parking Area</desc>
-    <sym>Parking Area</sym>
-    <type>Waypoint|Parking Area</type>
-    <cmt>Park here and walk 200m south.</cmt>
-  </wpt>
-</gpx>
-"""
 
 # ── Fixtures ──────────────────────────────────────────────────────────────────
 
-@pytest.fixture(scope="module")
-def tmp_db(tmp_path_factory):
-    db_path = tmp_path_factory.mktemp("data") / "test_importer.db"
-    init_db(db_path=db_path)
-    return db_path
-
-
 @pytest.fixture
 def gpx_file(tmp_path) -> Path:
-    f = tmp_path / "test.gpx"
-    f.write_text(SAMPLE_GPX, encoding="utf-8")
-    return f
+    return write_gpx(tmp_path, "test.gpx", SAMPLE_GPX)
 
 
 @pytest.fixture
 def wpts_file(tmp_path) -> Path:
-    f = tmp_path / "test-wpts.gpx"
-    f.write_text(SAMPLE_WPTS_GPX, encoding="utf-8")
-    return f
+    return write_gpx(tmp_path, "test-wpts.gpx", SAMPLE_WPTS_GPX)
 
 
 @pytest.fixture
 def zip_file(tmp_path, gpx_file, wpts_file) -> Path:
-    z = tmp_path / "test_pq.zip"
-    with zipfile.ZipFile(z, "w") as zf:
-        zf.write(gpx_file,  "test.gpx")
-        zf.write(wpts_file, "test-wpts.gpx")
-    return z
+    return make_zip(tmp_path, "test_pq.zip", {
+        "test.gpx": gpx_file,
+        "test-wpts.gpx": wpts_file,
+    })
 
 
 @pytest.fixture
 def multi_gpx_zip(tmp_path) -> Path:
-    z = tmp_path / "multi_pq.zip"
-    
-    # Content for a second unique GPX — log IDs must also differ because
-    # logs.log_id has a global UNIQUE constraint across all caches.
-    second_gpx = (SAMPLE_GPX
-        .replace("GC12345", "GCABCDE")
-        .replace("GC99999", "GCFGHIJ")
-        .replace('id="111"', 'id="333"')
-        .replace('id="112"', 'id="444"'))
-    
-    gpx1 = tmp_path / "first.gpx"
-    gpx1.write_text(SAMPLE_GPX, encoding="utf-8")
-    
-    gpx2 = tmp_path / "second.gpx"
-    gpx2.write_text(second_gpx, encoding="utf-8")
+    return make_zip(tmp_path, "multi_pq.zip", {
+        "first.gpx": SAMPLE_GPX,
+        "second.gpx": make_variant_gpx(),
+    })
 
-    with zipfile.ZipFile(z, "w") as zf:
-        zf.write(gpx1, "first.gpx")
-        zf.write(gpx2, "second.gpx")
-    return z
 
 # ── Basic import tests ────────────────────────────────────────────────────────
 
@@ -259,7 +138,6 @@ def test_import_with_companion_wpts(tmp_db, gpx_file, wpts_file):
 
 def test_import_zip(tmp_db, zip_file):
     """Verify that a PQ zip file is imported correctly end-to-end."""
-    # Reset DB for clean test — must delete children before parents (FK constraints)
     with get_session() as s:
         for cache in s.query(Cache).all():
             s.delete(cache)
@@ -323,15 +201,9 @@ def test_reimport_updates_not_duplicates(tmp_db, gpx_file):
 
 def test_import_empty_gpx(tmp_db, tmp_path):
     """A GPX file with no <wpt> elements should import cleanly with 0 results."""
-    empty = tmp_path / "empty.gpx"
-    empty.write_text(
-        '<?xml version="1.0" encoding="utf-8"?>'
-        '<gpx xmlns="http://www.topografix.com/GPX/1/0" version="1.0">'
-        '<n>Empty</n></gpx>',
-        encoding="utf-8",
-    )
+    f = write_gpx(tmp_path, "empty.gpx", EMPTY_GPX)
     with get_session() as s:
-        result = import_gpx(empty, s)
+        result = import_gpx(f, s)
     assert result.total == 0
     assert result.errors == []
 
@@ -347,11 +219,7 @@ def test_import_corrupt_gpx(tmp_db, tmp_path):
 
 def test_import_zip_empty(tmp_db, tmp_path):
     """A zip with no GPX files should return an error."""
-    z = tmp_path / "empty.zip"
-    txt = tmp_path / "readme.txt"
-    txt.write_text("no gpx here")
-    with zipfile.ZipFile(z, "w") as zf:
-        zf.write(txt, "readme.txt")
+    z = make_zip(tmp_path, "empty.zip", {"readme.txt": "no gpx here"})
     result = import_zip(z)
     assert result.total == 0
     assert len(result.errors) > 0
@@ -363,27 +231,11 @@ def test_import_zip_multiple_with_companion_wpts(tmp_db, tmp_path):
         for cache in s.query(Cache).all():
             s.delete(cache)
 
-    wpts_for_first = SAMPLE_WPTS_GPX  # links to GC12345 via suffix "2345"
-
-    second_gpx = (SAMPLE_GPX
-        .replace("GC12345", "GCABCDE")
-        .replace("GC99999", "GCFGHIJ")
-        .replace('id="111"', 'id="555"')
-        .replace('id="112"', 'id="666"'))
-
-    gpx1 = tmp_path / "first.gpx"
-    gpx1.write_text(SAMPLE_GPX, encoding="utf-8")
-    wpts1 = tmp_path / "first-wpts.gpx"
-    wpts1.write_text(wpts_for_first, encoding="utf-8")
-
-    gpx2 = tmp_path / "second.gpx"
-    gpx2.write_text(second_gpx, encoding="utf-8")
-
-    z = tmp_path / "multi_wpts.zip"
-    with zipfile.ZipFile(z, "w") as zf:
-        zf.write(gpx1, "first.gpx")
-        zf.write(wpts1, "first-wpts.gpx")
-        zf.write(gpx2, "second.gpx")
+    z = make_zip(tmp_path, "multi_wpts.zip", {
+        "first.gpx": SAMPLE_GPX,
+        "first-wpts.gpx": SAMPLE_WPTS_GPX,
+        "second.gpx": make_variant_gpx(log1="555", log2="666"),
+    })
 
     result = import_zip(z)
 
@@ -403,21 +255,7 @@ def test_import_gpx_inline_extra_waypoints(tmp_db, tmp_path):
         for cache in s.query(Cache).all():
             s.delete(cache)
 
-    gpx_with_inline_wpt = SAMPLE_GPX.replace(
-        "</gpx>",
-        '  <wpt lat="55.6770" lon="12.5690">\n'
-        '    <n>PK12345</n>\n'
-        '    <desc>Parking Area</desc>\n'
-        '    <sym>Parking Area</sym>\n'
-        '    <type>Waypoint|Parking Area</type>\n'
-        '    <cmt>Street parking available.</cmt>\n'
-        '  </wpt>\n'
-        '</gpx>'
-    )
-
-    f = tmp_path / "inline_wpts.gpx"
-    f.write_text(gpx_with_inline_wpt, encoding="utf-8")
-
+    f = write_gpx(tmp_path, "inline_wpts.gpx", make_gpx_with_inline_wpt())
     result = import_gpx(f)
 
     assert result.total == 2
